+++ CHANGELOG.md
@@ -5,10 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [5.0.0] - 2021-12-23
+### Breaking Changes:
+- **SlashCreator:** Removed `require-all` dependency and replaced `registerCommandsIn` options to only use strings
 ### Added:
 - **SlashCreator:** Add client passthrough option
-### Changed:
-- **SlashCreator:** Removed `require-all` dependency and replaced `registerCommandsIn` options to only use strings
 ### Removed:
 - **types:** Removed `RequireAllOptions` type
 ## [4.4.2] - 2021-12-20
@@ -360,7 +361,7 @@ This release features mostly completed documentation and changes to the packages
 ## [0.1.0] - 2020-12-15
 - Initial release.
 
-[Unreleased]: https://github.com/Snazzah/slash-create/compare/v4.4.2...HEAD
+[Unreleased]: https://github.com/Snazzah/slash-create/compare/v5.0.0...HEAD
 [0.1.0]: https://github.com/Snazzah/slash-create/releases/tag/v0.1.0
 [0.2.0]: https://github.com/Snazzah/slash-create/compare/v0.1.0...v0.2.0
 [1.0.0]: https://github.com/Snazzah/slash-create/compare/v0.2.0...v1.0.0
@@ -400,3 +401,4 @@ This release features mostly completed documentation and changes to the packages
 [4.4.0]: https://github.com/Snazzah/slash-create/compare/v4.3.0...v4.4.0
 [4.4.1]: https://github.com/Snazzah/slash-create/compare/v4.4.0...v4.4.1
 [4.4.2]: https://github.com/Snazzah/slash-create/compare/v4.4.1...v4.4.2
+[5.0.0]: https://github.com/Snazzah/slash-create/compare/v4.4.2...v5.0.0